requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
PyYAML>=6.0
tqdm>=4.65.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0
pycocotools>=2.0.6
geojson>=2.5.0
rasterio>=1.2.0
scipy>=1.7.0
shapely>=1.7.1 